+++ _NetcrackerHotel/src/main/java/by/netcracker/hotel/dao/UserDAO.java
@@ -4,10 +4,10 @@ import by.netcracker.hotel.entities.User;
 
 public interface UserDAO {
 
-	public void createUser(User user);
+    void createUser(User user);
 
-	public User readUser(User user);
+    User readUser(User user);
 
-	public void updateUser(User user);
+    void updateUser(User user);
 
 }
