+++ _NetcrackerHotel/src/main/webapp/WEB-INF/views/search_page.jsp
